+++ Assets/Demo/DemoPicker.cs
@@ -1,15 +1,27 @@
-using LoneTower.SRP;
-using System.Collections;
-using System.Collections.Generic;
+using LoneTower.SRP; 
 using UnityEngine;
 
 public class DemoPicker : MonoBehaviour {
 
-	[SRPDefaultBrush]
-	public DemoComponent singlePick;
+	//[SRPDefaultBrush]
+	//public DemoComponent singlePick;
+
+	//[SRPSingle]
+	//public DemoComponent component;
+
+	//[SRPPath]
+	//public PickableList<DemoComponent> pathPick;
+
+
+	//[SRPTopDownSingle]
+	//public DemoComponent topDownSingle;
 
 	[SRPTopDown]
-	public PickableList<DemoComponent> pathPick;
+	public PickableList<DemoComponent> topDownasdfas;
+
+	//[SRP("LogicPath", "ParserTransform", "DrawerPath", "ComponentPicker", "ComponentSerializer")]
+	//public PickableList<DemoComponent> pick;
+
 }