+++ Assets/SRP/Editor/SRPTypeParser.cs
@@ -5,7 +5,7 @@ using System.Reflection;
 using UnityEngine;
 namespace LoneTower.SRP {
 	public class SRPTypeParser {
-		 
+
 		public Type logic;
 		public Type parser;
 		public Type drawer;
@@ -14,30 +14,35 @@ namespace LoneTower.SRP {
 		public Type serializer;
 
 		public SRPTypeParser(string logic, string parser, string drawer, string picker, string serializer) {
-			this.logic = TryType(logic, typeof(LogicBase), typeof(LogicMain));
+			this.logic = TryType(logic, typeof(BrushBase), typeof(BrushMain));
 			this.parser = TryType(parser, typeof(ParserBase), typeof(ParserTransform));
 			this.drawer = TryType(drawer, typeof(DrawerBase), typeof(DrawerGeneric));
 			this.sceneInput = TryType(picker, typeof(ScenePickerBase), typeof(ComponentPicker));
 			this.serializer = TryType(serializer, typeof(SerializerBase), typeof(ComponentSerializer));
 		}
+
 		public SRPTypeParser(SRPAttribute attr) {
-			this.logic = TryType(attr.logic, typeof(LogicBase), typeof(LogicMain));
+			this.logic = TryType(attr.brush, typeof(BrushBase), typeof(BrushMain));
 			this.parser = TryType(attr.parser, typeof(ParserBase), typeof(ParserTransform));
 			this.drawer = TryType(attr.drawer, typeof(DrawerBase), typeof(DrawerGeneric));
 			this.sceneInput = TryType(attr.picker, typeof(ScenePickerBase), typeof(ComponentPicker));
 			this.serializer = TryType(attr.serializer, typeof(SerializerBase), typeof(ComponentSerializer));
 			this.selectType = attr.selectType;
+
 		}
+
 		Type TryType(string objType, Type expected, Type fallBack) {
-			Type current = Type.GetType("LoneTower.SRP." + objType);
+			objType = "LoneTower.SRP." + objType;
+			Type current = Type.GetType(objType);
 
 			if(CheckType(current, expected)) {
 				return current;
 			} else {
-				Debug.LogWarning($"[SRP] Wrong type {current}. Expected {expected}. Falling back to: {fallBack}");
+				Debug.LogWarning($"<b><color=#ED1E79>[SRP]</color></b> Wrong type <color=#4ec9b0>{objType}</color>. Expected <color=#4ec9b0>{expected.Name}</color>. Falling back to: <color=#4ec9b0>{fallBack.Name} </color>");
 				return fallBack;
 			}
 		}
+
 		static bool CheckType(Type t, Type g) {
 			if(t == null)
 				return false;