+++ Assets/SRP/Editor/Settings/SRPSettingsWindow.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
 
@@ -7,37 +5,59 @@ namespace LoneTower.SRP {
 	public class SRPSettingsWindow : EditorWindow {
 		[MenuItem("Window/SRP settings")]
 		static void Init() {
-			SRPSettingsWindow window = (SRPSettingsWindow)EditorWindow.GetWindow(typeof(SRPSettingsWindow));
+			SRPSettingsWindow window = (SRPSettingsWindow)EditorWindow.GetWindow<SRPSettingsWindow>("SRP Settings");
 			window.Show();
+			SRPSettings.Instance.Load();
 			window.colors = SRPSettings.colors;
 			ScriptableObject target = window;
 			SerializedObject so = new SerializedObject(target);
 			window.colorCollection = so.FindProperty("colors");
 
 		}
+
 		SerializedProperty colorCollection;
 		public Color[] colors;
+		bool foldMark = true;
+		bool foldDrawers = true;
+		Vector2 scrollPos = Vector2.zero;
+
 		void OnGUI() {
 			EditorGUI.BeginChangeCheck();
-			Color c = EditorGUILayout.ColorField(SRPSettings.choiceColor);
+			Color markColor = SRPSettings.MarkColor;
+			float scale = SRPSettings.Scale;
+			float markScale = SRPSettings.MarkScale;
+			float lineScale = SRPSettings.LineScale;
 
-			float scale = EditorGUILayout.FloatField("Scale", SRPSettings.Scale);
-			float lineScale = EditorGUILayout.FloatField("Line thickness", SRPSettings.LineScale);
-			float choiceScale = EditorGUILayout.FloatField("choice Size", SRPSettings.ChoiceSize);
-			EditorGUILayout.PropertyField(colorCollection, true);
+			if(foldMark = EditorGUILayout.Foldout(foldMark, "Mark")) {
+				EditorGUI.indentLevel++;
+				markColor = EditorGUILayout.ColorField("color", SRPSettings.MarkColor);
+				markScale = EditorGUILayout.FloatField("choice Size", SRPSettings.MarkScale);
+				EditorGUI.indentLevel--;
+			}
+
+			if(foldDrawers = EditorGUILayout.Foldout(foldDrawers, "Drawer")) {
+				EditorGUI.indentLevel++;
+				scale = EditorGUILayout.FloatField("Scale", SRPSettings.Scale);
+				lineScale = EditorGUILayout.FloatField("Line thickness", SRPSettings.LineScale);
+				EditorGUI.indentLevel--;
+			}
+			;
+			using(var scrollView = new EditorGUILayout.ScrollViewScope(scrollPos)) {
+				scrollPos = scrollView.scrollPosition;
+				EditorGUILayout.PropertyField(colorCollection, true);
+			}
 
 
 			if(EditorGUI.EndChangeCheck()) {
 				SRPSettings.colors = colors;
-				SRPSettings.choiceColor = c;
+				SRPSettings.MarkColor = markColor;
 				SRPSettings.Scale = scale;
-				SRPSettings.ChoiceSize = choiceScale;
+				SRPSettings.MarkScale = markScale;
 				SRPSettings.LineScale = lineScale;
 
 				SRPSettings.Instance.Save();
+				DrawerBase.ColorBank = new SelectionBank<Color>(SRPSettings.colors);
 			}
-			EditorGUILayout.LabelField("Changes will be applied on recompile");
-			//so.ApplyModifiedProperties(); // Remember to apply modified properties
 
 		}
 	}