+++ Assets/SRP/Runtime/SRPAttribute.cs
@@ -9,7 +9,7 @@ namespace LoneTower.SRP {
 	[AttributeUsage(AttributeTargets.Field, AllowMultiple = true, Inherited = true)]
 	public class SRPAttribute : PropertyAttribute {
 
-		public string logic;
+		public string brush;
 		public string parser;
 		public string drawer;
 		public string picker;
@@ -17,22 +17,21 @@ namespace LoneTower.SRP {
 
 
 		public Type selectType;
-		 
-		public SRPAttribute(string logic, string parser, string drawer, string sceneInput, string serializer) {
-			this.logic = logic;
+
+		public SRPAttribute(string brush, string parser, string drawer, string sceneInput, string serializer) {
+			this.brush = brush;
 			this.parser = parser;
 			this.drawer = drawer;
 			this.picker = sceneInput;
 			this.serializer = serializer;
-		} 
+		}
 		public SRPAttribute(params string[] data) {
-			this.logic = data[0];
+			this.brush = data[0];
 			this.parser = data[1];
 			this.drawer = data[2];
 			this.picker = data[3];
 			this.serializer = data[4];
 		}
-
 
 	}