+++ Assets/SRP/Runtime/SRPBrushes.cs
@@ -6,32 +6,29 @@ using UnityEngine;
 namespace LoneTower.SRP {
 
 	public class SRPDefaultBrush : SRPAttribute {
-		public SRPDefaultBrush() : base("LogicMain", "ParserTransform", "DrawerGeneric", "ComponentPicker", "ComponentSerializer") {
+		public SRPDefaultBrush() : base("BrushMain", "ParserTransform", "DrawerGeneric", "ComponentPicker", "ComponentSerializer") {
 
 		}
 	}
 	public class SRPPath : SRPAttribute {
-		public SRPPath() : base("LogicPath", "ParserTransform", "DrawerPath", "ComponentPicker", "ComponentSerializer") {
+		public SRPPath() : base("BrushPath", "ParserTransform", "DrawerPath", "ComponentPicker", "ComponentSerializer") {
 
 		}
 	}
-	public class SRPPathSingle : SRPAttribute {
-		public SRPPathSingle() : base("LogicSingle", "ParserTransform", "DrawerPath", "ComponentPicker", "ComponentSerializer") {
 
-		}
-	}
+
 	public class SRPSingle : SRPAttribute {
-		public SRPSingle() : base("LogicSingle", "ParserTransform", "DrawerGeneric", "ComponentPicker", "ComponentSerializer") {
+		public SRPSingle() : base("BrushSingle", "ParserTransform", "DrawerGeneric", "ComponentPicker", "ComponentSerializer") {
 
 		}
 	}
 	public class SRPTopDown : SRPAttribute {
-		public SRPTopDown() : base("LogicMain", "ParserTransform", "DrawerTopDown", "ComponentPicker", "ComponentSerializer") {
+		public SRPTopDown() : base("BrushMain", "ParserTransform", "DrawerTopDown", "ComponentPicker", "ComponentSerializer") {
 
 		}
 	}
 	public class SRPTopDownSingle : SRPAttribute {
-		public SRPTopDownSingle() : base("LogicSingle", "ParserTransform", "DrawerTopDown", "ComponentPicker", "ComponentSerializer") {
+		public SRPTopDownSingle() : base("BrushSingle", "ParserTransform", "DrawerTopDown", "ComponentPicker", "ComponentSerializer") {
 		}
 	}
 }